chore: Remove everything related to server and build FilterNode
